requirements.txt
# Core dependencies
openai>=1.3.0
mistralai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Document processing
PyMuPDF>=1.23.0
python-docx>=1.0.0

# Image processing
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

